requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyMySQL==0.10.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
WTForms==2.3.3
blinker==1.4
click==7.1.2
Flask==1.1.2
Flask-Admin==1.5.6
Flask-GoogleCharts==0.1.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysqlclient==2.0.1
pyChart.JS==0.3.0
PyMySQL==0.10.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
WTForms==2.3.3
